Refactor typehint function
